requirements.txt
Babel==2.5.3
beautifulsoup4==4.6.0
boto3==1.7.28
botocore==1.10.28
brotlipy==0.7.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.4.2
croniter==0.3.20
Django==2.0.9
django-anymail==5.0
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-filter==1.1.0
django-fsm==2.6.0
django-htmlmin==0.10.0
django-ipware==2.0.1
django-money==0.12.3
django-phonenumber-field==2.0.0
django-redis-cache==1.7.1
django-redis-sessions==0.6.1
django-rq==1.0.1
django-storages==1.6.6
django-tenants==2.0.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djangorestframework-msgpack==1.0.2
docutils==0.14
gunicorn==19.7.1
html5lib==1.0.1
idna==2.6
jmespath==0.9.3
msgpack-python==0.5.6
olefile==0.44
phonenumberslite==8.9.2
Pillow==5.0.0
psycopg2==2.7.3.2
psycopg2-binary==2.7.4
py-moneyed==0.7.0
pycparser==2.18
PyJWT==1.6.1
python-dateutil==2.7.1
pytz==2018.3
q==2.6
raven==6.8.0
redis==2.10.6
requests==2.21.0
rq==0.10.0
rq-scheduler==0.8.2
s3transfer==0.1.13
six==1.11.0
sorl-thumbnail==12.4.1
sqlparse==0.2.4
urllib3==1.24.1
webencodings==0.5.1
whitenoise==3.3.1
